+++ Src/GLWorld.App/shader/Shader.h
@@ -13,7 +13,7 @@ public:
 	void use();
 	void setBool(const std::string& name, bool value) const;
 	void setInt(const std::string& name, int value) const;
-	void setFloat(const std::string& name, int value) const;
+	void setFloat(const std::string& name, float value) const;
 	void setFloat3(const std::string& name, float x, float y, float z) const;
 
 	~Shader()
